Refactor stop_words into its own module
