+++ tests/granular/test_stop_words.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from nlp_profiler.constants import NaN
-from nlp_profiler.core import gather_stop_words, count_stop_words  # noqa
+from nlp_profiler.stop_words import gather_stop_words, count_stop_words  # noqa
 
 text_with_a_number = '2833047 people live in this area'
